+++ src/main/java/com/paoloamosso/piggus/PiggusApplication.java
@@ -28,4 +28,6 @@ public class PiggusApplication extends SpringBootServletInitializer {
 
 		SpringApplication.run(PiggusApplication.class, args);
 	}
+
+
 }